requirements/base.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.12.4
python-dotenv==1.1.1

numpy==2.3.5
faiss-cpu==1.13.0
rank-bm25==0.2.2

openai==2.8.1
tiktoken==0.7.0
sentence-transformers==5.1.2
torch==2.9.1
transformers==4.57.1

pdfplumber==0.11.8
streamlit==1.51.0
